Remove unused code
and some small code cleanup
