+++ src/fheroes2/castle/castle_well.cpp
@@ -35,20 +35,6 @@
 #include "text.h"
 #include "world.h"
 
-struct dwelling_t : public std::pair<u32, u32>
-{
-    dwelling_t( u32 type, u32 count )
-        : std::pair<u32, u32>( type, count ){};
-};
-
-struct dwellings_t : public std::vector<dwelling_t>
-{
-    dwellings_t()
-    {
-        reserve( 6 );
-    };
-};
-
 u32 HowManyRecruitMonster( const Castle & castle, u32 dw, const Funds & add, Funds & res )
 {
     const Monster ms( castle.GetRace(), castle.GetActualDwelling( dw ) );
