+++ src/fheroes2/castle/mageguild.cpp
@@ -28,7 +28,6 @@
 #include "rand.h"
 #include "settings.h"
 
-Spell GetUniqueCombatSpellCompatibility( const SpellStorage &, int race, int level );
 Spell GetCombatSpellCompatibility( int race, int level );
 Spell GetUniqueSpellCompatibility( const SpellStorage & spells, const int race, const int level );
 Spell GetGuaranteedDamageSpellForMageGuild();
@@ -107,14 +106,6 @@ void MageGuild::educateHero( HeroBase & hero, int guildLevel, bool hasLibrary ) 
         // this method will check wisdom requirement
         hero.AppendSpellsToBook( MageGuild::GetSpells( guildLevel, hasLibrary ) );
     }
-}
-
-Spell GetUniqueCombatSpellCompatibility( const SpellStorage & spells, int race, int lvl )
-{
-    Spell spell = GetCombatSpellCompatibility( race, lvl );
-    while ( spells.isPresentSpell( spell ) )
-        spell = GetCombatSpellCompatibility( race, lvl );
-    return spell;
 }
 
 Spell GetUniqueSpellCompatibility( const SpellStorage & spells, const int race, const int lvl )
