+++ src/fheroes2/kingdom/race.cpp
@@ -74,22 +74,3 @@ int Race::Rand( void )
 
     return Race::NECR;
 }
-
-int Race::FromInt( int race )
-{
-    switch ( race ) {
-    case KNGT:
-    case BARB:
-    case SORC:
-    case WRLK:
-    case WZRD:
-    case NECR:
-    case MULT:
-    case RAND:
-        return race;
-    default:
-        break;
-    }
-
-    return Race::NONE;
-}
