+++ src/fheroes2/maps/maps.cpp
@@ -623,11 +623,6 @@ StreamBase & operator>>( StreamBase & sb, IndexObject & st )
     return sb >> st.first >> st.second;
 }
 
-StreamBase & operator>>( StreamBase & sb, IndexDistance & st )
-{
-    return sb >> st.first >> st.second;
-}
-
 StreamBase & operator>>( StreamBase & sb, ObjectColor & st )
 {
     return sb >> st.first >> st.second;
